Use more precise rounding for autopy's pixel derivations
The int conversion has a floor-like behavior where a width of
984.6153846153845 points and 1.04 scale will result in 1024.0
pixels but a height of 738.4615384615383 points and 1.04 scale
will result in 767.9999999999999 pixels with floor of 767 or
1 pixel less than the actual height. The mouse location also
has point units whose conversion to pixels could be more precise.

Use `round()` instead of `int()` to round the width and height,
as well as mouse location more precisely. Since the rounding may
now slightly overshoot the allowed maximum size in points, let's
preserve the partial pixels in the internal width as well height
representations and use this safe maximum when taking screenshots
together with a simpler minimal of 1x1 point.