Work around display controller test race conditions
It seems that the VNCDoTool display controller could fail to click
or type properly (maybe because it is a network protocol based tool?)
and this could result in a race with the Qt6 app windows not closing.
Add a retry functionality to support testing of less stable controllers
like this (with the suggestion to make them more reliable upstream).

While at it, reuse more python library functionality for the Qt6
app closing checks and clean the image logs in a reliable way if
a test is retried (in advance to avoid skipped cleanup for failures).

No more need for longer app closing timeout - we can use the default.