Unify the temporary file name use for windows compatibility
The temp file created by python cannot be removed on windows for
certain paths (which are not even selected by us) then written on.
Instead of adding delete flags to prevent more restrictive permissions
on windows let's update the temporary file name handling in the autopy
finder to coincide with the logic in autopy and other controllers as
well as all other locations in the code base.

One different to pick on is to make an extra effort at preventing
resource leak which is not done (and likely not needed) elsewhere.