Use pixel-relevant screen size with the AutoPy display controller
The screen we are capturing must be the same pixel size as the
typical needles we have provide. The reason this would have worked
before is that previous test environments had much smaller AutoPy
scale factor leading to a match of the searched image despite slighly
larger or smaller haystack version. Testing on MacOS with a scale
factor of 2.0 or greater no longer allows that.

Resize the scaled image captured by AutoPy instead to guarantee the
pixel screen size.

All of this more or less drops the need of using the scale factor
in any display controller operations but it is still kept around
for information if needed (as it is an important autopy parameter).