Perform identical DPI and scaling for Qt6 apps for all OS-es
Use environment variables for the fixed DPI settings that should
make pixel counts independent of the environment screen.

Fix zero margins and spacing for the Qt6 image window but adapt
the locations once while taking the other possible approach for
the Qt6 app - keep the prevous locations but fix the margins and
spacing to match the previous ones for certain default displays.
Zero spacing and magins was too ugly for the Qt6 application but
not that relevant for the simpler Qt6 image. Also use fixed and
more universal fonts in a more standard way (not stylesheet).

Additionally sort the Qt imports better and only load the Qt app
if running as a script.