Skip known failing tests on macOS
This comprises mostly of tests not finding the green box in the CI
(but finding it in a non-CI macOS environment as well as environment
with identical resolution) which need more debugging as well as
problems with AutoPy handling the SHIFT key on macOS, drag/drop
problems, and mouse-up and middle-click problems.

Also standardize the platform checks for Windows in comparison.